EcommerceFullStackProject/ai_service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sentence-transformers==3.3.1
faiss-cpu==1.12.0
numpy==2.2.1
pymongo==4.10.1
requests==2.32.5
